agents/requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
requests>=2.31.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
mysql-connector-python>=8.3.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
PyGithub>=2.1.1
tenacity>=8.2.3
schedule>=1.2.1
python-telegram-bot>=21.0
aiosqlite>=0.20.0
# aiosqlite must be installed separately if not pulled in by python-telegram-bot
